refactor: racingGame domain 추가
