+++ src/main/java/racingcar/controller/GameController.java
@@ -2,6 +2,7 @@ package racingcar.controller;
 
 import racingcar.model.Car;
 import racingcar.model.Cars;
+import racingcar.model.RacingGame;
 import racingcar.util.InputValidator;
 import racingcar.view.InputView;
 import racingcar.view.OutputView;
@@ -10,12 +11,7 @@ import java.util.*;
 import java.util.stream.Collectors;
 
 public class GameController {
-    private static final String COMMA = ",";
-    private static final String BLANK = "";
-    private static final String SPACE = " ";
-
-    private Cars cars;
-    private int trialNum;
+    private RacingGame racingGame;
 
     public void run() {
         initGame();
@@ -25,46 +21,20 @@ public class GameController {
 
     private void initGame() {
         String carNames = InputView.inputCarNames();
-        List<String> splitCarNames = splitCars(carNames);
-        cars = initCars(splitCarNames);
         String inputTrialNum = InputView.inputTrialNum();
-        trialNum = convertToInteger(inputTrialNum);
+        racingGame = new RacingGame(carNames, inputTrialNum);
     }
 
-    public List<String> splitCars(String carNames) {
-        InputValidator.validateNameInput(carNames);
-        List<String> splitCarNames = Arrays.asList(splitNameWithoutSpace(carNames));
-        InputValidator.validateEmptyCarName(splitCarNames);
-        InputValidator.validateCarNameLength(splitCarNames);
-        InputValidator.validateDuplicateName(splitCarNames);
-
-        return splitCarNames;
-    }
-
-    private String[] splitNameWithoutSpace(String carNames) {
-        return carNames.replaceAll(SPACE, BLANK).split(COMMA);
-    }
-
-    public Cars initCars(List<String> carNames) {
-        return new Cars(carNames.stream()
-                .map(Car::new)
-                .collect(Collectors.toList()));
-    }
-
-    private int convertToInteger(String number) {
-        InputValidator.validateTrialInput(number);
-        return Integer.parseInt(number);
-    }
 
     private void runGame() {
         OutputView.printTrialResult();
-        for (int i = 0; i < trialNum; i++) {
-            cars.race();
-            OutputView.printCarPosition(cars);
+        for (int i = 0; i < racingGame.getTrialNum(); i++) {
+            racingGame.race();
+            OutputView.printCarPosition(racingGame.getCars());
         }
     }
 
     private void endGame() {
-        OutputView.printWinnerName(cars.findWinner());
+        OutputView.printWinnerName(racingGame.getCars().findWinner());
     }
 }
